django seminar2 user profile property added & bit styling
