+++ seminar/accounts/templates/accounts/signup.html
@@ -17,5 +17,16 @@
     <input name="password2" type="password" value="">
     <br>
     <br>
+    학교:
+    <br>
+    <input name="college" type="text" value="">
+    <br>
+    <br>
+    전공:
+    <br>
+    <input name="major" type="text" value="">
+    <br>
+    <br>
     <input type="submit" value="Sign Up!">
+
 </form>
