+++ seminar/feedpage/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import Feed, Profile
+from .models import Feed
 
 # Register your models here.
 admin.site.register(Feed)
-admin.site.register(Profile)
