+++ seminar/feedpage/models.py
@@ -18,20 +18,3 @@ class Feed(models.Model): # 모델 클래스명은 단수형을 사용 (Feeds(x)
 
     def __str__(self):
         return self.title   
-
-class Profile(models.Model):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    college = models.CharField(max_length=20, blank=True)
-    major = models.CharField(max_length=20, blank=True)
-
-    def __str__(self):  
-        return 'id=%d, user_id=%d, college=%s, major=%s' % (self.id, self.user.id, self.college, self.major)
-
-@receiver(post_save, sender=User)
-def create_user_profile(sender, instance, created, **kwargs):  
-    if created:
-        Profile.objects.create(user=instance)
-
-@receiver(post_save, sender=User)
-def save_user_profile(sender, instance, **kwargs):  
-    instance.profile.save()
