+++ seminar/feedpage/templates/feedpage/index.html
@@ -1,10 +1,14 @@
 {% extends 'base.html' %}
 
 {% block content %}
-    <h1>feed page</h1>
-    <a href="/feeds/new">new feed</a> 
+    <div style="margin: 10px;">
+      <h1>feed page</h1>
+      <a href="/feeds/new"
+         style="background-color: #ffb03b; border-radius: 10px; width: auto; height: auto; padding: 5px; color: #ffffff;">
+         new feed</a><br>
+    </div>
     {% for feed in feeds %} 
-      <div style="border: 1px solid grey;">
+      <div style="border: 2px dashed #ffb03b; margin: 10px; padding: 5px; width: 500px; display: inline-block">
         <a href="/feeds/{{ feed.id }}">더보기</a>
         <a href="/feeds/{{ feed.id }}/delete" onclick="return confirm('정말 삭제하시겠습니까?')">삭제</a>
         <p>{{ feed.title }}</p>
